feat: file structure 수정
