+++ backend/app/main.py
@@ -1,0 +1,12 @@
+from fastapi import FastAPI 
+from .routers import articles
+
+app = FastAPI() 
+
+
+app.include_router(articles.router)
+
+
+@app.get("/")
+async def root():
+    return {"Greetings" : "Hello, World"}
